Avoid using immediately-invoked functions factories
It complicates reading the code
