+++ packages/slate-editor/src/modules/editor/plugins/withDeserializeHtml/createDeserialize/index.ts
@@ -1,1 +1,1 @@
-export { createDeserialize } from './createDeserialize';
+export { createDeserializer } from './createDeserializer';
